+++ src/functions.php
@@ -6,12 +6,14 @@ if (! function_exists('take')) {
     /**
      * Create a new piped item from a given value.
      *
-     * @param mixed $value
+     * @param mixed $value The value you want to process.
+     * @param string $identifier The identifier to replace the value with in method calls that
+     *     don't take the value as first parameter.
      *
      * @return \SebastiaanLuca\PipeOperator\Item
      */
-    function take($value) : Item
+    function take($value, $identifier = '$$') : Item
     {
-        return new Item($value);
+        return new Item($value, $identifier);
     }
 }