+++ CHANGELOG.md
@@ -13,3 +13,4 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Added custom pipe operator
 - Allow using direct method calls instead of using pipe operator
 - Add shorthand take method
+- Allow setting custom pipe operator